rename TyGrit.scene to TyGrit.belief_state
Future-extension rename: the module previously held a single point-cloud
scene representation, but it will grow to host other belief
representations alongside it. Update all import sites accordingly.
