--- TyGrit/scene/__init__.py
@@ -1,7 +1,0 @@
-"""Scene — the world model / belief state."""
-
-from TyGrit.scene.config import PointCloudSceneConfig, SceneConfig
-from TyGrit.scene.pointcloud_scene import PointCloudScene
-from TyGrit.scene.scene import Scene
-
-__all__ = ["Scene", "SceneConfig", "PointCloudScene", "PointCloudSceneConfig"]
